Review comment: move Crypto back to distributed package
